Optimize tag manage ui

+++ pages/api/tag/index.ts
@@ -8,7 +8,15 @@ type Data = {
   name: string
 }
 const getTags = async (req: NextApiRequest) => {
-  const data = await prisma.tag.findMany()
+  const params = req.query
+  const searchKey = params?.search_key || ''
+  const data = await prisma.tag.findMany({
+    where: {
+      name: {
+        contains: searchKey as string,
+      },
+    },
+  })
   return { msg: '', code: 0, data: data }
 }
 const postTags = async (req: NextApiRequest) => {
@@ -33,10 +41,11 @@ const putTags = async (req: NextApiRequest) => {
   })
 }
 const deleteTag = async (req: NextApiRequest) => {
-  const params = req.body
+  const { id } = req.query
+  if (!id) return { msg: '', code: 100045, data: '缺少参数' }
   await prisma.tag.delete({
     where: {
-      id: params.id,
+      id: +id,
     },
   })
   return { msg: '', code: 0, data: '' }
@@ -56,10 +65,6 @@ export default async function handler(
   }
   if (req.method === 'PUT') {
     const r = await putTags(req)
-    res.status(200).send(r as any)
-  }
-  if (req.method === 'DELETE') {
-    const r = await deleteTag(req)
     res.status(200).send(r as any)
   }
 }
